fix: RLP deserialization of big.Int(0) was different.
Due to the design of RLP, 0x80 is originally treated as both an empty string or int = 0.
In standard RLP, if [0x80] is found, Stream.Bytes() interprets it as an empty string and will pass [] to decodeBigInt().
If an aeternity '0' (which is [0x00]) is found, Stream.Bytes() will pass [0x00] to decodeBigInt().
Here two things will happen: [0x00] is detected as a leading zero (not true in this case) and big.Int.SetBytes([0x00]) is called, which is not the same as big.Int.SetBytes([]), even though they are the same value of 0.
This commit changes the leading zero detection and has a special if clause to pass in a [] to Int.SetBytes() if b = [0x00]